+++ apps/business-management/src/tokens/get-one-token/get-one-token.query.ts
@@ -4,6 +4,8 @@
  *
  * Licensed under BSD 3-Clause License. See LICENSE for terms.
  */
-export * from './members'
-export * from './organizations'
-export * from './tokens'
+import { OrganizationQuery } from '@turnly/core'
+
+export class GetOneTokenQuery extends OrganizationQuery {
+  public readonly secret: string
+}