+++ packages/grpc/src/consumers/business-management/tokens/index.ts
@@ -4,6 +4,5 @@
  *
  * Licensed under BSD 3-Clause License. See LICENSE for terms.
  */
-export * from './members'
-export * from './organizations'
-export * from './tokens'
+export * from './tokens.proxy'
+export * from './tokens.types'